requirements.txt
pandas
numpy
matplotlib
requests
Flask
Flask-Cors
boto3
textract-trp
